backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
daphne==4.0.0
setuptools==82.0.1
